# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MiniProyecto",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api"
        }
    ],
    "paths": {
...
        "/enterprises": {
            "get": {
                "tags": [
                    "Enterprises"
                ],
                "summary": "Create enterprise",
                "description": "Create and return one enterprise",
                "operationId": "getEnterprises",
                "responses": {
                    "200": {
                        "description": "Todas las empresas",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Enterprises"
                ],
                "summary": "Create enterprise",
                "description": "Create and return one enterprise",
                "operationId": "create_enterprise",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "address",
                                    "logo",
                                    "phone",
                                    "dni",
                                    "coin",
                                    "description"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "address": {
                                        "type": "string"
                                    },
                                    "logo": {
                                        "format": "byte"
                                    },
                                    "phone": {
                                        "type": "string"
                                    },
                                    "dni": {
                                        "type": "string"
                                    },
                                    "coin": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "Pepe",
                                    "address": "Murcia 24/70 Malecon - Habana",
                                    "logo": "base64....",
                                    "phone": "+1522469335",
                                    "dni": "001458796",
                                    "coin": "USD",
                                    "description": "Hola esto es una descripcion"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Empresa registrada satisfactoriamente",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/enterprises/{enterpriseId}": {
            "put": {
                "tags": [
                    "Enterprises"
                ],
                "summary": "Create enterprise",
                "description": "Update and return one enterprise",
                "operationId": "update_enterprise",
                "parameters": [
                    {
                        "name": "enterpriseId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "address",
                                    "logo",
                                    "phone",
                                    "dni",
                                    "coin",
                                    "description"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "address": {
                                        "type": "string"
                                    },
                                    "logo": {
                                        "format": "byte"
                                    },
                                    "phone": {
                                        "type": "string"
                                    },
                                    "dni": {
                                        "type": "string"
                                    },
                                    "coin": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "Pepe",
                                    "address": "Murcia 24/70 Malecon - Habana",
                                    "logo": "base64....",
                                    "phone": "+1522469335",
                                    "dni": "001458796",
                                    "coin": "USD",
                                    "description": "Hola esto es una descripcion"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Empresa registrada satisfactoriamente",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Enterprises"
                ],
                "summary": "Delete enterprise",
                "description": "Delete one enterprise",
                "operationId": "delete_enterprise",
                "parameters": [
                    {
                        "name": "enterpriseId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Empresa registrada satisfactoriamente",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}